Update retreive to validate pageNum
retrieve() now validates the pageNum passed through as a param to the over-ride function.
We also check if the instance has been destroyed before calling the over-ride. Finally, moved the validation logic of the pageNum into the code block that handles triggering the over-ride since it has no other relevance to the retrieve() function.