Update _binding() to use .on() and .off()
jQuery has deprecated .bind() and .unbind() in version 3.0. Its recommended to update to use .on() and .off() which have been around since jQuery v1.7. All reference to .bind() and .unbind() have been update and the _binding() function refactored to support TRUE / FALSE in addition to maintaining backwards compatibility for the moment with 'bind' and 'unbind' values.